Use same generation logic for pipelines and for testing locally; also adds news generation support for testing locally.
