+++ src/generator/create-folder.js
@@ -3,7 +3,10 @@ import * as fs from 'fs';
 export default async function createFolder(folderPath) {
   try {
     await fs.mkdirSync(folderPath, { recursive: true });
-  } catch (error) {
-    throw new Error(error);
+  }
+  catch (error) {
+    if (error.code !== 'EEXIST') {
+      throw new Error(error);
+    }
   }
 }
