+++ src/generator/generate-local.js
@@ -1,189 +1,136 @@
 import * as fs from 'fs';
-
 import fetch from 'node-fetch';
-import nunjucks from 'nunjucks';
-import { minify } from 'html-minifier';
 
-import { NunjucksLoader } from './nunjucks-loader';
-import removeFolder from './remove-folder';
-import createFolder from './create-folder';
 import asyncForEach from './async-foreach';
+import createFolder from './create-folder';
+import generate from './generate';
 import getAsset from './get-asset';
-import mapPages from './map-pages';
-import storeFiles from './store-files';
 import rateLimiter from './rate-limiter';
-
+import removeFolder from './remove-folder';
+import storeFiles from './store-files';
 
 const cwd = process.cwd();
+const designSystemPath = `${cwd}/node_modules/@ons/design-system`;
 const buildDestination = `${cwd}/dist`;
-const viewsPath = `${cwd}/src/views`;
 
 const languages = ['en', 'cy', 'ni'];
 
 const localPort = 4040;
-const enSite = 'http://en.localhost:' + localPort + '/';
-const cySite = 'http://cy.localhost:' + localPort + '/';
+process.env.EN_SITE = 'http://en.localhost:' + localPort + '/';
+process.env.CY_SITE = 'http://cy.localhost:' + localPort + '/';
 
 const assetFetchConcurrencyLimit = 50;
-const designSystemPath = `${cwd}/node_modules/@ons/design-system`;
-const searchPaths = [viewsPath, `${viewsPath}/templates`, `${designSystemPath}`];
-const nunjucksLoader = new NunjucksLoader(searchPaths);
-const nunjucksEnvironment = new nunjucks.Environment(nunjucksLoader);
-
-nunjucks.configure(null, {
-  watch: false,
-  autoescape: true
-});
 
 const statusParam = process.env.CONTENT_STATUS;
 
 let apiURL = 'http://localhost/api';
-let assetURL = 'http://localhost/assets/uploads/';
+let assetBaseUrl = 'http://localhost/assets/uploads/';
 
-
-async function getContent() {
-  
+async function getSourceData() {
   const requests = languages.map(async language => {
-    
     try {
-      let entriesJson, globalsJson, assetsJson;
+      let entriesJson, newsJson, globalsJson, newsSettingsJson, assetsJson;
 
       console.log(`Fetching ${language} entries`);
       const entriesResponse = await fetch(`${apiURL}/entries-en.json?status=${statusParam}`);
       if (entriesResponse.status === 500) {
         throw new Error('Error fetching entries: ' + entriesResponse.status);
-      } else {
+      }
+      else {
         console.log(`Fetching entries ${language} complete`);
+      }
+
+      console.log(`Fetching ${language} news`);
+      const newsResponse = await fetch(`${apiURL}/news-${language}.json`);
+      if (newsResponse.status === 500) {
+        throw new Error('Error fetching news: ' + newsResponse.status);
+      }
+      else {
+        console.log(`Fetching news ${language} complete`);
       }
 
       console.log(`Fetching ${language} globals`);
       const globalsResponse = await fetch(`${apiURL}/globals-${language}.json`);
       if (globalsResponse.status === 500) {
         throw new Error('Error fetching globals: ' + globalsResponse.status);
-      } else {
+      }
+      else {
         console.log(`Fetching globals ${language} complete`);
+      }
+
+      console.log(`Fetching ${language} news settings from news globals`);
+      const newsSettingsResponse = await fetch(`${apiURL}/news-globals-${language}.json`);
+      if (newsSettingsResponse.status === 500) {
+        throw new Error('Error fetching globals: ' + newsSettingsResponse.status);
+      }
+      else {
+        console.log(`Fetching news settings from news globals ${language} complete`);
       }
 
       console.log(`Fetching assets`);
       const assetsResponse = await fetch(`${apiURL}/assets.json`);
       if (assetsResponse.status === 500) {
         throw new Error('Error fetching assets: ' + assetsResponse.status);
-      } else {
+      }
+      else {
         console.log(`Fetching assets complete`);
       }
 
       entriesJson = await entriesResponse.json();
+      newsJson = await newsResponse.json();
+      newsSettingsJson = await newsSettingsResponse.json();
       globalsJson = await globalsResponse.json();
       assetsJson = await assetsResponse.json();
+
+      newsSettingsJson.data[0].featuredEntry = entriesJson.data.find(entry => entry.id === newsSettingsJson.data[0].featuredEntry);
 
       await removeFolder(buildDestination);
 
       return {
         pages: entriesJson.data,
+        news: newsJson.data[0],
         globals: globalsJson.data[0],
-        assets: assetsJson.data
+        newsSettings: newsSettingsJson.data[0],
+        assets: assetsJson.data,
       };
-    } catch (error) {
+    }
+    catch (error) {
       console.log(error);
       process.exit(1);
     }
   });
 
-  const data = await Promise.all(requests);
+  return await Promise.all(requests);
+}
 
-  await createFolder(buildDestination);
-
+async function getSourceAssets(sourceData) {
   await asyncForEach(languages, async (language, index) => {
-    const mappedPages = mapPages(data[index].pages, data[index].globals, enSite, cySite);
-    renderSite(language, mappedPages);
-
-    await rateLimiter(data[index].assets, async asset => await getAssets(language, asset), assetFetchConcurrencyLimit);
+    await rateLimiter(sourceData[index].assets, async asset => await getAssets(language, asset), assetFetchConcurrencyLimit);
     await storeFiles(designSystemPath, buildDestination, language);
-  });
-
-}
-
-async function renderSite(key, pages) {
-  const siteFolder = `${buildDestination}/${key}`;
-  asyncForEach(pages, page => renderPage(siteFolder, page));
-}
-
-function renderPage(siteFolder, page) {
-  return new Promise(resolve => {
-    nunjucks.compile(`{% extends "${page.type}.html" %}`, nunjucksEnvironment).render(page, async (error, result) => {
-      
-      if (error) {
-        throw new Error(error);
-        process.exit(1);
-        
-      }
-
-      const folderPath = page.url ? `${siteFolder}/${page.url}` : siteFolder;
-
-      await createFolder(folderPath);
-
-      const html = minify(result, {
-        removeComments: true,
-        collapseWhitespace: true
-      });
-
-      await fs.writeFileSync(`${folderPath}/index.html`, html, 'utf8');
-
-      resolve();
-    });
-  });
-}
-
-function storeAsset(asset, sourceDes, buildDes) {
-  return new Promise(resolve => {
-    const url = sourceDes + asset.url;
-
-    getAsset(url)
-      .then(async data => {
-        fs.writeFileSync(`${buildDes}/${asset.url}`, data);
-        
-        resolve();
-      })
-      .catch(error => {
-        throw new Error(error);
-        process.exit(1);
-      });
-  
   });
 }
 
 async function getAssets(key, asset) {
-
-    const url = assetURL;
-  
-    if (asset.transforms && asset.transforms.length) {
-
-      asset.transforms.forEach(async(transform) => {
-        const folderPath = `${buildDestination}/${key}/${transform.location}`;
-        createFolder(folderPath);
-        const buildDes = `${buildDestination}/${key}/${transform.location}`;
-        const sourceDes = `${url}/${transform.location}/`;
-        await storeAsset(asset, sourceDes, buildDes);
-      })
-    
-    } 
-    
-    const buildDes = `${buildDestination}/${key}`;
-    const sourceDes = `${url}`;
-    await storeAsset(asset, sourceDes, buildDes);
-    
-
+  if (asset.transforms && asset.transforms.length) {
+    asset.transforms.forEach(async(transform) => {
+      const folderPath = `${buildDestination}/${key}/${transform.location}`;
+      createFolder(folderPath);
+      const buildDes = `${buildDestination}/${key}/${transform.location}`;
+      const sourceDes = `${assetBaseUrl}${transform.location}/`;
+      const url = sourceDes + asset.url;
+      try {
+        const data = await getAsset(url);
+        fs.writeFileSync(`${buildDes}/${asset.url}`, data);
+      }
+      catch {
+        console.warn(`Could not fetch asset: ${url}`);
+      }
+    })
+  } 
 }
 
-async function run() {
-  getContent();
-}
-
-run();
-
-
-
-nunjucksEnvironment.addFilter('setAttr', function(dictionary, key, value) {
-  dictionary[key] = value;
-  return dictionary;
-});
+(async () => {
+  const sourceData = await getSourceData();
+  await getSourceAssets(sourceData);
+  await generate(sourceData, languages, buildDestination);
+})();
